feat: teach bloom_elem to return a double
Now that we have bloom_approximate_fp_rate, we can respond to member-
ship tests with 0 if the item is not in the set, or 1-fp_rate if it
might be.

This effectively responds with the percentage probability that the
tested element is in the set, allowing the caller to see the filter's
reliability decrease with each inserted item.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>